requirements.txt
django
django-cors-headers
groq
faiss-cpu
sentence-transformers
numpy
gunicorn
scikit-learn
requests